creating case documents to its own component
